+++ src/api/tests/feature_tests/steps/endpoint_lambda_mapping.py
@@ -25,9 +25,11 @@ def get_endpoint_lambda_mapping() -> ENDPOINT_LAMBDA_MAPPING:
     import api.createCpmProductForEpr.index
     import api.createDevice.index
     import api.createProductTeam.index
+    import api.deleteCpmProduct.index
     import api.readCpmProduct.index
     import api.readDevice.index
     import api.readProductTeam.index
+    import api.searchCpmProduct.index
     import api.searchDevice.index
     import api.status.index
 
@@ -44,9 +46,13 @@ def get_endpoint_lambda_mapping() -> ENDPOINT_LAMBDA_MAPPING:
             "Device?device_type={device_type}&foo={foo}": api.searchDevice.index,
             "Device?foo={foo}": api.searchDevice.index,
             "ProductTeam/{product_team_id}/Product/{product_id}": api.readCpmProduct.index,
+            "ProductTeam/{product_team_id}/Product": api.searchCpmProduct.index,
             "ProductTeam/{product_team_id}": api.readProductTeam.index,
             # "Organization/{id}": api.readProductTeam.index,
             "_status": api.status.index,
+        },
+        "DELETE": {
+            "ProductTeam/{product_team_id}/Product/{product_id}": api.deleteCpmProduct.index,
         },
     }