+++ src/layers/domain/common_steps/tests/test_read_product.py
@@ -1,0 +1,81 @@
+from uuid import uuid4
+
+import pytest
+from domain.common_steps.read_product import before_steps
+from domain.core.cpm_product.v1 import CpmProduct
+from domain.core.root.v3 import Root
+from domain.repository.cpm_product_repository.v3 import CpmProductRepository
+from domain.repository.errors import ItemNotFound
+from domain.repository.product_team_repository.v2 import ProductTeamRepository
+from domain.response.validation_errors import InboundValidationError
+from event.aws.client import dynamodb_client
+from event.step_chain import StepChain
+
+from test_helpers.dynamodb import mock_table
+
+TABLE_NAME = "my-table"
+
+
+@pytest.mark.parametrize(
+    ["event", "expected_exception"],
+    [
+        (
+            {"pathParameters": {"product_team_id": "does-not-exist"}},
+            InboundValidationError,
+        ),
+        (
+            {
+                "pathParameters": {
+                    "product_team_id": "does-not-exist",
+                    "product_id": "does-not-exist",
+                },
+            },
+            ItemNotFound,
+        ),
+    ],
+)
+def test_read_product_steps_bad_input(event: dict, expected_exception: type[Exception]):
+    step_chain = StepChain(step_chain=before_steps)
+
+    mocked_cache = {"DYNAMODB_CLIENT": dynamodb_client(), "DYNAMODB_TABLE": TABLE_NAME}
+    with mock_table(table_name=TABLE_NAME):
+        step_chain.run(init=event, cache=mocked_cache)
+
+    assert isinstance(step_chain.result, expected_exception)
+
+
+def test_read_product_steps_good_input():
+    product_team_id = uuid4()
+    ods_code = "AAA"
+    step_chain = StepChain(step_chain=before_steps)
+
+    mocked_cache = {"DYNAMODB_CLIENT": dynamodb_client(), "DYNAMODB_TABLE": TABLE_NAME}
+    with mock_table(table_name=TABLE_NAME):
+        product_team_repo = ProductTeamRepository(
+            table_name=mocked_cache["DYNAMODB_TABLE"],
+            dynamodb_client=mocked_cache["DYNAMODB_CLIENT"],
+        )
+        org = Root.create_ods_organisation(ods_code=ods_code)
+        product_team = org.create_product_team(id=product_team_id, name="foo-team")
+        product_team_repo.write(product_team)
+
+        product_repo = CpmProductRepository(
+            table_name=mocked_cache["DYNAMODB_TABLE"],
+            dynamodb_client=mocked_cache["DYNAMODB_CLIENT"],
+        )
+        product = product_team.create_cpm_product(name="foo-product")
+        product_repo.write(product)
+
+        step_chain.run(
+            init={
+                "pathParameters": {
+                    "product_team_id": str(product_team_id),
+                    "product_id": str(product.id),
+                },
+            },
+            cache=mocked_cache,
+        )
+
+    assert isinstance(step_chain.result, CpmProduct)
+    assert step_chain.result.product_team_id == product_team_id
+    assert step_chain.result.ods_code == ods_code